Initial commit - Sports Club REST API

+++ src/main/java/com/example/sportsclubapi/repository/SportRepository.java
@@ -3,5 +3,5 @@ package com.example.sportsclubapi.repository;
 import com.example.sportsclubapi.model.Sport;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SportRepository extends JpaRepository<Sport, Integer> {
+public interface SportRepository extends JpaRepository<Sport, Long> {
 }
